Return `Result` in accessors instead of panic (#286)
As previously suggested in #274, I replaced the `.expect` with returning
a `Result` in the operation accessors of ODS generated dialects.

I also added some variants to the `Error` enum to support this. The
`Infallible` variant was added to allow using `TryInto` to convert
`Attribute` into itself, such that we avoid needing to handle
`Attribute` differently from `StringAttribute`, `IntAttribute`, etc. But
if you prefer, I can change the macro code to deal with this case
instead, I'm honestly not a big fan of my `Infallible` hack either.

---------

Co-authored-by: Yota Toyama <raviqqe@gmail.com>