Refactor OperationField (#287)
Moved most `OperationField` fields to specific variants of the
`FieldKind` enum and moved `param_type` and `return_type` to methods
instead of determining them when constructing the `OperationField`.

I'm also refactoring the `Operation::from_def` method for a later PR,
since it's way to large and hard to read. I'm replacing the
`VariadicKind` updating with a more idiomatic `VariadicKindIter` that
implements `Iterator`.