Refactor `Operation::from_def` (#288)
I've made the following changes in this PR:

- Split `Operation::from_def` in several `collect_` methods
- Instead of collecting all operation fields into a single `Vec`,
collect them into several smaller `Vec`s and chain the iterators
on-demand in the `fields` function. This makes it a bit easier to handle
errors.
- Replaced `initial_variadic_kind` and `update_variadic_kind` closures
with a `VariadicKindIter` such that we can `zip` this iterator with the
iterators for operands and results. I'm not sure if this is the best
solution, but I do personally prefer it over my previous solution.

There are still some things I'm unsure about, and would like to hear
your thoughts. I will add some comments to the code diff.

Edit: I'm sorry for the messy commit log, but the diff of all changes
should still be clean and it shouldn't matter as long as you squash the
commits.

---------

Co-authored-by: Yota Toyama <raviqqe@gmail.com>